refactor(ui): consolidate Send and Sweep precondition alerts
